Added fix for full sentence quotes
Added a fix for quotes in "" that spans a full paragraph:

"The lazy developer jumped over the try-catch-pattern."

Caveat: If you start an article with something in "" that is not a quote, it might be changed to chevrons:

"Lazy" is a great way to describe this developer.  -->  »Lazy« is a great way to describe this developer.

Or you could run into an issue, if you start a paragraph with a double quote and include a punctuation mark:

"Lazy." It's the new big thing among developers. --> »Lazy.« It's the big thing among developers.

Further testing needed, but basic functionality is there.